requirements.txt
Flask==3.1.0
flask_cors==5.0.1
mysql_connector_python==9.2.0
nltk==3.9.1
numpy==2.2.4
torch==2.6.0
